ZOMATO RESTAURANT SUCCESS FACTORS ANALYSIS
Exploratory Data Analysis (EDA) serves as an initial phase in the realm of Machine Learning, extensively employed to gain a profound understanding of the dataset. While not mandatory for model development, EDA is strongly recommended as it facilitates a nuanced comprehension of the data. When executed meticulously, EDA unveils insights that may not be readily apparent.

What can we do with this Dataset

- Knowing basic composition of data
- Removing duplicates
- Dealing with missing values
- Understanding features
- Plotting horizontal bar charts (multicolor)
- Using groupby, apply, and unique functions
- Scatter plot
- Word Cloud
- Box plot
- Density plot
- Bar Charts
- Drawing insights and conclusions from data